+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.css";
-import Home from "./components/Home";
+import Home from "./components/Home/Home";
 
 const App = () => {
   return (
